requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
duckdb
pandas==2.3.2
python-multipart==0.0.6

# PostgreSQL dependencies (solo asyncpg, sin psycopg2)
asyncpg
sqlalchemy

# Environment variables
python-dotenv